refactor : 참가자 추상 클래스 추가
